move run test case to issue
